changed the data types correctly

+++ header/header.h
@@ -18,10 +18,10 @@ typedef long long llong;
 ulong hamming_weight (ulong u);
 ulong pow2 (int power);
 int sign (ulong e);
-ulong * allocate_table ( ulong m );
-long * allocate_long_table ( ulong m );
+ulong * allocate_table (ulong m);
+long * allocate_long_table (ulong m);
 ulong scalar_product (ulong *table_f1, ulong *table_f2, ulong m);
 ulong integer_product (ulong vector1, ulong vector2);
-int * walsh_transform (int *f, int m);
+long * walsh_transform (long *f, ulong m);
 ulong table_hamming_weight (ulong *val, ulong m);
 #endif
